+++ helpers/getImage.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var _ = require('lodash');
+const SafeString = require('handlebars').SafeString;
 const common = require('./../lib/common');
 
 function helper(paper) {
@@ -23,17 +24,23 @@ function helper(paper) {
             // If preset is one of the given presets in _images
             width = parseInt(presets[presetName].width, 10) || 5120;
             height = parseInt(presets[presetName].height, 10) || 5120;
-            size = width + 'x' + height;
-
+            size = `${width}x${height}`;
         } else if (sizeRegex.test(settings[presetName])) {
             // If preset name is a setting and match the NNNxNNN format
             size = settings[presetName];
+            width = parseInt(size.split('x')[0], 10);
+            height = parseInt(size.split('x')[1], 10);
         } else {
             // Use the original image size
             size = 'original';
         }
 
-        return image.data.replace('{:size}', size);
+        if (Number.isInteger(image.width) && Number.isInteger(image.height)
+            && Number.isInteger(width) && Number.isInteger(height)) {
+            size = `${Math.min(image.width, width)}x${Math.min(image.height, height)}`
+        }
+
+        return new SafeString(image.data.replace('{:size}', size));
     });
 };